Seperate out cookie helper/consent logic
