+++ source/assets/javascripts/cookie_consent.js
@@ -1,43 +1,32 @@
-// Used to set, get and check for cookie consent
-function checkForCookies(name) {
-    return !!getCookie(name);
-}
-
-function getCookie(name) {
-    let cookieArray = document.cookie.split(';');
-    return cookieArray.find(cookieCrumb => cookieCrumb.substring(name.length, 0) === name);
-}
-
-function setCookie(name, value, expiry) {
-    let date = new Date();
-    date.setTime(date.getTime() + (expiry * 24 * 60 * 60 * 1000));
-    let expires = "expires=" + date.toUTCString();
-    document.cookie = name + "=" + value + ";" + expires + ";path=/; sameSite=none; Secure";
-}
+import { checkForCookies, setCookie } from './cookie_helper'
 
 function updateBanner() {
-    document.getElementById("cookieActionsContainer").style.display = "none";
-    document.getElementById("cookieConfirmation").style.display = "block";
+  document.getElementById("cookieActionsContainer").style.display = "none";
+  document.getElementById("cookieConfirmation").style.display = "block";
 }
 
 function unhideBanner() {
-    document.getElementById("appCookieBanner").style.display = "block";
+  document.getElementById("appCookieBanner").style.display = "block";
 }
 
 function hideBanner() {
-    document.getElementById("appCookieBanner").style.display = "none";
+  document.getElementById("appCookieBanner").style.display = "none";
 }
 
 function acceptAllCookies() {
-    setCookie("ghre_allow_cookies", true, 365);
-    updateBanner();
+  setCookie("ghre_allow_cookies", true, 365);
+  updateBanner();
 }
 
-if(!checkForCookies("ghre_allow_cookies")) {
-    unhideBanner(); // unhide cookie banner if consent not given
+function cookieConsentGiven() {
+  return checkForCookies("ghre_allow_cookies");
+}
+
+if (!cookieConsentGiven()) {
+  unhideBanner(); // unhide cookie banner if consent not given
 };
 
 document.getElementById("acceptCookies").onclick = acceptAllCookies;
 document.getElementById("hideButton").onclick = hideBanner;
 
-export { setCookie, getCookie, checkForCookies, acceptAllCookies, hideBanner, unhideBanner };
+export { acceptAllCookies, hideBanner, unhideBanner };
